Fix product combinations/features with nginx
prestashop at least since 1.7.2.2 performs queries to URLs such as
https://shop.com/adminXXX/index.php/feature/get-feature-values/9 in order to
retrieve AJAX values for features and attributes. Without this, the nginx
configuration always returned an error 404 and combinations broke.
So we need to add the "feature" module as well to the list of symfony
controllers that are handed to the adminXXX/index.php ...

Symptoms of this error are:
1) Product attributes cannot be selected from predefined values on the product
  details page and
2) When you attempt to generate combinations of a product, you get a green
  "Changes saved" popup, but no combinations are actually created.