Fix product combinations/features with nginx
prestashop at least since 1.7.2.2 performs queries to URLs such as
https://shop.com/adminXXX/index.php/attribute/product-form-images/11 in order to
retrieve AJAX values for features and attributes. Without this, the nginx
configuration always returned an error 404 and combinations broke.
So we need to add the "attribute" module as well to the list of symfony
controllers that are handed to the adminXXX/index.php ...

In contrast to the missing "feature" module, there were no user-visible failures,